backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain
langchain-ollama
langgraph
langgraph-checkpoint
langgraph-checkpoint-redis
qdrant-client
redis
tavily-python
pypdf
python-multipart
httpx
pydantic-settings
black
isort
flake8
numpy